Rename DatabaseObjectComparator class to be DatabaseObjectCollectionComparator for clarity DAT-10112 (#3544)
* Added ability to specify a storedlogic comparator

DAT-10112

* [DAT-10112] Removes use of getDbObjectComparator when finding missing objects.

* Renamed class DatabaseObjectComparator to DatabaseObjectCollectionComparator

DAT-10112

* Another push to branch DAT-10112
