+++ liquibase-core/src/main/java/liquibase/serializer/core/yaml/YamlSnapshotSerializer.java
@@ -1,6 +1,7 @@
 package liquibase.serializer.core.yaml;
 
 import liquibase.GlobalConfiguration;
+import liquibase.diff.compare.DatabaseObjectCollectionComparator;
 import liquibase.exception.UnexpectedLiquibaseException;
 import liquibase.serializer.LiquibaseSerializable;
 import liquibase.serializer.SnapshotSerializer;
@@ -10,7 +11,6 @@ import liquibase.statement.SequenceCurrentValueFunction;
 import liquibase.statement.SequenceNextValueFunction;
 import liquibase.structure.DatabaseObject;
 import liquibase.structure.DatabaseObjectCollection;
-import liquibase.structure.DatabaseObjectComparator;
 import liquibase.structure.core.Column;
 import liquibase.util.ISODateFormat;
 import liquibase.util.StringUtil;
@@ -86,7 +86,7 @@ public class YamlSnapshotSerializer extends YamlSerializer implements SnapshotSe
             SortedMap<String, Object> returnMap = new TreeMap<>();
             for (Map.Entry<Class<? extends DatabaseObject>, Set<? extends DatabaseObject>> entry : ((DatabaseObjectCollection) object).toMap().entrySet()) {
                 ArrayList value = new ArrayList(entry.getValue());
-                Collections.sort(value, new DatabaseObjectComparator());
+                Collections.sort(value, new DatabaseObjectCollectionComparator());
                 returnMap.put(entry.getKey().getName(), value);
             }
             return returnMap;
